+++ realm/realm-library/src/androidTest/java/io/realm/TestHelper.java
@@ -46,6 +46,11 @@ import java.util.concurrent.TimeUnit;
 import io.realm.entities.AnnotationIndexTypes;
 import io.realm.entities.AllTypes;
 import io.realm.entities.NullTypes;
+import io.realm.entities.PrimaryKeyAsBoxedByte;
+import io.realm.entities.PrimaryKeyAsBoxedInteger;
+import io.realm.entities.PrimaryKeyAsBoxedLong;
+import io.realm.entities.PrimaryKeyAsBoxedShort;
+import io.realm.entities.PrimaryKeyAsString;
 import io.realm.entities.StringOnly;
 import io.realm.internal.Table;
 import io.realm.internal.TableOrView;
@@ -383,6 +388,171 @@ public class TestHelper {
         }
 
         return config.build();
+    }
+
+    /**
+     * Adds a String type PrimaryKey object to a realm with values for name field (PrimaryKey) and id field
+     */
+    public static PrimaryKeyAsString addStringPrimaryKeyObjectToTestRealm(Realm testRealm, String primaryFieldValue, long secondaryFieldValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsString obj = new PrimaryKeyAsString();
+        obj.setName(primaryFieldValue);
+        obj.setId(secondaryFieldValue);
+        testRealm.copyToRealm(obj);
+        testRealm.commitTransaction();
+        return obj;
+    }
+
+    /**
+     * Populates a realm with String type Primarykey objects for a number of numberOfPopulation - 1,
+     * starting with iteratorBeginValue. One object is setup to have given values from parameters.
+     */
+    public static void populateTestRealmWithStringPrimaryKey(Realm testRealm, String primaryFieldValue, long secondaryFieldValue, int numberOfPopulation, int iteratorBeginValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsString userObj = new PrimaryKeyAsString();
+        userObj.setName(primaryFieldValue);
+        userObj.setId(secondaryFieldValue);
+        testRealm.copyToRealm(userObj);
+        int idValue = iteratorBeginValue;
+        for (int i = 0; i < numberOfPopulation - 1; ++i, ++idValue) {
+            PrimaryKeyAsString obj = new PrimaryKeyAsString();
+            obj.setName(String.valueOf(idValue));
+            obj.setId(idValue);
+            testRealm.copyToRealm(obj);
+        }
+        testRealm.commitTransaction();
+    }
+
+    /**
+     * Adds a Byte type PrimaryKey object to a realm with values for id field (PrimaryKey) and name field
+     */
+    public static PrimaryKeyAsBoxedByte addBytePrimaryKeyObjectToTestRealm(Realm testRealm, Byte primaryFieldValue, String secondaryFieldValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedByte obj = new PrimaryKeyAsBoxedByte();
+        obj.setId(primaryFieldValue);
+        obj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(obj);
+        testRealm.commitTransaction();
+        return obj;
+    }
+
+    /**
+     * Populates a realm with Byte type Primarykey objects for a number of numberOfPopulation - 1,
+     * starting with iteratorBeginValue. One object is setup to have given values from parameters.
+     */
+    public static void populateTestRealmWithBytePrimaryKey(Realm testRealm, Byte primaryFieldValue, String secondaryFieldValue, int numberOfPopulation, int iteratorBeginValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedByte userObj = new PrimaryKeyAsBoxedByte();
+        userObj.setId(primaryFieldValue);
+        userObj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(userObj);
+        byte idValue = (byte)iteratorBeginValue;
+        for (int i = 0; i < numberOfPopulation - 1; ++i, ++idValue) {
+            PrimaryKeyAsBoxedByte obj = new PrimaryKeyAsBoxedByte();
+            obj.setId(new Byte(idValue));
+            obj.setName(String.valueOf(idValue));
+            testRealm.copyToRealm(obj);
+        }
+        testRealm.commitTransaction();
+    }
+
+    /**
+     * Adds a Short type PrimaryKey object to a realm with values for id field (PrimaryKey) and name field
+     */
+    public static PrimaryKeyAsBoxedShort addShortPrimaryKeyObjectToTestRealm(Realm testRealm, Short primaryFieldValue, String secondaryFieldValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedShort obj = new PrimaryKeyAsBoxedShort();
+        obj.setId(primaryFieldValue);
+        obj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(obj);
+        testRealm.commitTransaction();
+        return obj;
+    }
+
+    /**
+     * Populates a realm with Short type Primarykey objects for a number of numberOfPopulation - 1,
+     * starting with iteratorBeginValue. One object is setup to have given values from parameters.
+     */
+    public static void populateTestRealmWithShortPrimaryKey(Realm testRealm, Short primaryFieldValue, String secondaryFieldValue, int numberOfPopulation, int iteratorBeginValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedShort userObj = new PrimaryKeyAsBoxedShort();
+        userObj.setId(primaryFieldValue);
+        userObj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(userObj);
+        short idValue = (short)iteratorBeginValue;
+        for (int i = 0; i < numberOfPopulation - 1; ++i, ++idValue) {
+            PrimaryKeyAsBoxedShort obj = new PrimaryKeyAsBoxedShort();
+            obj.setId(new Short(idValue));
+            obj.setName(String.valueOf(idValue));
+            testRealm.copyToRealm(obj);
+        }
+        testRealm.commitTransaction();
+    }
+
+    /**
+     * Adds a Integer type PrimaryKey object to a realm with values for id field (PrimaryKey) and name field
+     */
+    public static PrimaryKeyAsBoxedInteger addIntegerPrimaryKeyObjectToTestRealm(Realm testRealm, Integer primaryFieldValue, String secondaryFieldValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedInteger obj = new PrimaryKeyAsBoxedInteger();
+        obj.setId(primaryFieldValue);
+        obj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(obj);
+        testRealm.commitTransaction();
+        return obj;
+    }
+
+    /**
+     * Populates a realm with Integer type Primarykey objects for a number of numberOfPopulation - 1,
+     * starting with iteratorBeginValue. One object is setup to have given values from parameters.
+     */
+    public static void populateTestRealmWithIntegerPrimaryKey(Realm testRealm, Integer primaryFieldValue, String secondaryFieldValue, int numberOfPopulation, int iteratorBeginValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedInteger userObj = new PrimaryKeyAsBoxedInteger();
+        userObj.setId(primaryFieldValue);
+        userObj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(userObj);
+        int idValue = iteratorBeginValue;
+        for (int i = 0; i < numberOfPopulation - 1; ++i, ++idValue) {
+            PrimaryKeyAsBoxedInteger obj = new PrimaryKeyAsBoxedInteger();
+            obj.setId(new Integer(idValue));
+            obj.setName(String.valueOf(idValue));
+            testRealm.copyToRealm(obj);
+        }
+        testRealm.commitTransaction();
+    }
+
+    /**
+     * Adds a Long type PrimaryKey object to a realm with values for id field (PrimaryKey) and name field
+     */
+    public static PrimaryKeyAsBoxedLong addLongPrimaryKeyObjectToTestRealm(Realm testRealm, Long primaryFieldValue, String secondaryFieldValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedLong obj = new PrimaryKeyAsBoxedLong();
+        obj.setId(primaryFieldValue);
+        obj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(obj);
+        testRealm.commitTransaction();
+        return obj;
+    }
+
+    /**
+     * Populates a realm with Long type Primarykey objects for a number of numberOfPopulation - 1,
+     * starting with iteratorBeginValue. One object is setup to have given values from parameters.
+     */
+    public static void populateTestRealmWithLongPrimaryKey(Realm testRealm, Long primaryFieldValue, String secondaryFieldValue, long numberOfPopulation, long iteratorBeginValue) {
+        testRealm.beginTransaction();
+        PrimaryKeyAsBoxedLong userObj = new PrimaryKeyAsBoxedLong();
+        userObj.setId(primaryFieldValue);
+        userObj.setName(secondaryFieldValue);
+        testRealm.copyToRealm(userObj);
+        long idValue = iteratorBeginValue;
+        for (long i = 0; i < numberOfPopulation - 1; ++i, ++idValue) {
+            PrimaryKeyAsBoxedLong obj = new PrimaryKeyAsBoxedLong();
+            obj.setId(new Long(idValue));
+            obj.setName(String.valueOf(idValue));
+            testRealm.copyToRealm(obj);
+        }
+        testRealm.commitTransaction();
     }
 
     public static void populateTestRealmForNullTests(Realm testRealm) {