+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.89.0
 
+### Breaking changes
+
+* @PrimaryKey field value can be null for String, Byte, Short, Integer, and Long types. Older Realms should be migrated, using RealmObjectSchema.setNullable(), or by adding the @Required annotation. (#2515).
+
 ### Enhancements
 
 * RealmObjectSchema.isPrimaryKey(String) (#2440)